refactor(novedades): cambio de clase por color y refactor de servicios
